refactor: rename isCrossRiver to hasCrossedRiver in Position.ts (#23)
* refactor(test): Replaced new Pawn() with createPawn() Pawn.test.ts

* refactor(logic): Use ternary operator in constructor in Board.ts

* chore: Added eslint-disable no-unused-vars comment in PieceType.ts

* chore: update eslint to exclude console.log statements in codebase

* refactor: isCrossRiver to hasCrossedRiver

"isCrossRiver" is not gramatically correct and it makes more sense to use a paste tense phrase "hasCrossedRiver" as it will be more intuitive when used as a boolean function

refs: https://dev.to/michi/tips-on-naming-boolean-variables-cleaner-code-35ig

* chore: removed unused import

* refactor(test): Change it() to test()

Reason for this is because the it() function is meant to be used semantically e.g. it("should be false"). Hence if the test case does not fit this "should do something" etc. , just use test() instead of it().
See link below for more examples.

refs: https://stackoverflow.com/questions/45778192/what-is-the-difference-between-it-and-test-in-jest

* refactor(test): Remove unnecessary "Test: " texts in describe() blocks

There is no need to include the "Test: " because you are already describing a test.

* docs(piece): add JSDoc-style comments for RookFactory.ts and Rook.ts

* test: update old isCrossRiver to hasCrossedRiver in Position.test.ts

---------

Co-authored-by: Wu Zengfu <67005197+wuzengfu@users.noreply.github.com>
Co-authored-by: Wu Zengfu <wuzengfu666@gmail.com>